+++ src/actions/stac/StacMap.js
@@ -1,0 +1,23 @@
+import StacActionPlugin from "../StacActionPlugin";
+import { URI } from 'stac-js/src/utils.js';
+import i18n from "../../i18n";
+import BIconMap from '~icons/bi/map';
+
+export default class StacMap extends StacActionPlugin {
+
+  get show() {
+    return this.object.isItem;
+  }
+
+  get uri() {
+    return URI('https://developmentseed.org/stac-map/').addQuery('href', this.object.getAbsoluteUrl());
+  }
+
+  get icon() {
+    return BIconMap;
+  }
+
+  get text() {
+    return i18n.global.t('actions.openIn', {service: 'stac-map'});
+  }
+}