+++ src/components/Item.vue
@@ -15,8 +15,10 @@
       <Keywords v-if="showKeywordsInItemCards && keywords.length > 0" :keywords="keywords" variant="primary" />
       <b-card-text><small class="datetime" v-html="displayTime" /></b-card-text>
     </b-card-body>
-    <b-card-footer v-if="$slots.footer">
-      <slot name="footer" :data="data" :source="item" />
+    <b-card-footer>
+      <slot name="footer" :data="data" :source="item">
+        <StacActions v-if="data" :data="data" variant="outline-primary" compact size="sm" />
+      </slot>
     </b-card-footer>
   </b-card>
 </template>
@@ -28,6 +30,7 @@ import { mapState, mapGetters } from 'vuex';
 import FileFormatsMixin from './FileFormatsMixin';
 import CardMixin from './CardMixin';
 import StacLink from './StacLink.vue';
+import StacActions from './StacActions.vue';
 import { STAC } from 'stac-js';
 import { formatTemporalExtent, formatTimestamp } from '@radiantearth/stac-fields/formatters';
 import { BCard, BCardBody, BCardFooter, BCardText, BCardTitle, BCardImg } from 'bootstrap-vue-next';
@@ -36,6 +39,7 @@ export default defineComponent({
   name: 'Item',
   components: {
     StacLink,
+    StacActions,
     BCard,
     BCardImg,
     BCardBody,